backend/requirements.txt
fastapi
uvicorn
google-genai
google-generativeai
sqlalchemy
email-validator
python-dotenv
pandas
passlib
PyJWT
